WIP: Mark all channels as read
