+++ plugins/chat/assets/javascripts/discourse/models/chat-channel.js
@@ -192,8 +192,11 @@ export default class ChatChannel extends RestModel {
       return;
     }
 
-    return ajax(`/chat/${this.id}/read/${messageId}.json`, {
+    // TODO (martin) Change this to use chatApi service once we change this
+    // class not to use RestModel
+    return ajax(`/chat/api/tracking/read/me`, {
       method: "PUT",
+      data: { message_id: messageId, channel_id: this.id },
     }).then(() => {
       this.currentUserMembership.last_read_message_id = messageId;
     });
