+++ plugins/chat/assets/javascripts/discourse/services/chat-channels-manager.js
@@ -85,8 +85,9 @@ export default class ChatChannelsManager extends Service {
 
   @debounce(300)
   async markAllChannelsRead() {
-    return this.chatApi.updateCurrentUserTracking().then(() => {
-      this.channels.forEach((channel) => {
+    return this.chatApi.updateCurrentUserTracking().then((response) => {
+      response.updated_memberships.forEach((membership) => {
+        let channel = this.channels.findBy("id", membership.chat_channel_id);
         channel.currentUserMembership.unread_count = 0;
         channel.currentUserMembership.unread_mentions = 0;
       });
